requirements.txt
argcomplete==1.12.3
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorlog==6.6.0
cryptography==36.0.1
databases==0.5.3
distlib==0.3.4
filelock==3.4.0
google-ads==14.1.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
httplib2==0.20.2
idna==3.3
mypy==0.930
mypy-extensions==0.4.3
mysqlclient==2.1.0
nox==2021.10.1
oauth2client==1.5.2
oauthlib==3.1.1
packaging==21.3
platformdirs==2.4.0
proto-plus==1.18.1
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.28
toml==0.10.2
tomli==2.0.0
types-python-dateutil==2.8.3
types-PyYAML==6.0.1
typesystem==0.3.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
